Implement sendRequest in Capwebct payment module and connector (to send VCR and HFMR and update accounts in the module) Handle a specific URL for each WS call in connector Add individual roles in the home folder XML

+++ ExternalModules/Payment/db/schema.pgsql.sql
@@ -247,7 +247,7 @@
         child_surname varchar(255),
         label varchar(255),
         unit_price int4,
-        quantity int4,
+        quantity numeric(19, 2),
         value int4,
         invoice_fk_id int8,
         invoice_detail_index int4,
